changed to double

+++ include/nextweek/box.cuh
@@ -47,7 +47,7 @@ public:
     }
   }
   __host__ __device__ void rotate_y(Hittable **&ss,
-                                    float angle) {
+                                    double angle) {
     for (int i = start_index; i < end_index; i++) {
       list[i] = new RotateY(list[i], angle);
       ss[i] = list[i];
